requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1748348059416/work
africastalking==1.2.9
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1736248061654/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1749972191589/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.2.1
dnspython==2.7.0
fastapi==0.115.13
filelock==3.13.1
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1739279199749/work
fsspec==2024.6.1
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1733299680490/work
google-pasta @ file:///home/conda/feedstock_root/build_artifacts/google-pasta_1733852499742/work
greenlet==3.2.3
grpcio @ file:///croot/grpc-split_1742490263735/work
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py @ file:///croot/h5py_1750862990877/work
hf-xet==1.1.5
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.33.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
Jinja2==3.1.4
joblib==1.5.1
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1747722781675/work
Mako==1.3.10
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1750360292101/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219583522/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
ml_dtypes @ file:///croot/ml_dtypes_1745327848793/work
mpi4py @ file:///croot/mpi4py_1749458488241/work
mpmath==1.3.0
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1748346368839/work
networkx==3.3
numpy==2.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
optree @ file:///croot/optree_1742210181724/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pillow==11.2.1
protobuf @ file:///croot/protobuf_1742419705443/work/bazel-bin/python/dist/protobuf-5.29.3-cp312-abi3-linux_x86_64.whl#sha256=b288453436273a46653983ade721c0c62cfa5ef6c012b110674c369d25536d88
psycopg2-binary==2.9.10
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pymongo==4.13.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1743371105/work/dist
safetensors==0.5.3
schema==0.7.7
scikit-learn==1.7.0
scipy==1.16.0
setuptools==78.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.3
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1733820672324/work/tensorboard-2.18.0-py3-none-any.whl#sha256=107ca4821745f73e2aefa02c50ff70a9b694f39f790b11e6f682f7d326745eab
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=162df89696269a70a88358ad14c08b73a057ccde1ec28f3dd314ca4fc4fd9ce4
tensorflow @ file:///croot/tensorflow-split_1746131914377/work/tensorflow_pkg/tensorflow-2.18.1-cp312-cp312-linux_x86_64.whl#sha256=d716f3b2e23b49d34b5cd3be2e79ab940718dfa60a1b7399b3213131b8d1e72d
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1746039537955/work
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1748959427/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
uvicorn==0.34.3
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wheel==0.45.1
wrapt @ file:///croot/wrapt_1736540904746/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard @ file:///croot/zstandard_1731356346222/work
